Refactoring and some improvements
