--- src/main/java/me/petrolingus/electricfieldlines/Algorithm.java
@@ -1,41 +1,0 @@
-package me.petrolingus.electricfieldlines;
-
-public class Algorithm {
-
-    public int n;
-
-    public double[][] matrixA;
-
-    public double[] vectorB;
-
-    public double[] vectorC;
-
-    public void init() {
-        matrixA = new double[n][n];
-        vectorB = new double[n];
-        vectorC = new double[n];
-    }
-
-    public void createMatrixA() {
-
-        for (int i = 0; i < n; i++) {
-            for (int j = 0; j < n; j++) {
-                if (i == j) {
-
-                } else {
-
-                }
-            }
-        }
-
-    }
-
-    public void createVectorB() {
-
-    }
-
-    public void findSolution() {
-
-    }
-
-}
